Finished movement controls for steamvr, narrator fixes
First things first: the joystick works great. One major downside of the joystick
on the windows MR controller however is that it's in your hand at an angle. That
makes it somewhat unclear which direction is up on the joystick. So it's pretty
easy to walk at a slant. This is fixed by a hardcoded arrow icon directly above
the joystick. It feels really good. The main bug I was struggling with before by
the way with movement was that I failed to convert euler angles into rads.

A bool has been added in PlayerMove to indicate when the player is moving
forward or backwards. This simplifies the event catching for
IntroductionNarrator.

Adjusted default controls to use the joystick configuration.