Bulk of refactoring, minus testing. Clarified PlayerMic purpose.
PlayerMic is responsible purely for listening to the record button and reacting
accordingly. It has public properties to let outside classes know when a
recording is in progress and when the record button is pressed. The latter is
important as it hides OVR/SteamVR details away from the other classes.
IntroductionNarrator and ObjectFocus are allowed to get the recording from
PlayerMic and play it in their respective AudioSources. They've been refactored
to have less dependency on PlayerMic callback functions and to instead keep
state themselves. ObjectFocus in particular underwent a major refactoring as
variables were deemed unnecessary and eliminated entirely. None of this is
tested yet. Note that behavior in IntroductionNarrator may be slightly tweaked
because I don't see any code that listens for the "listen" button. It seemed to
just play back the initial recording right away. Still does, but I'm going to
change it.