Refactor for C# convention, Mic/Audio Refactors
This change ended up being rather large. Let's break it down, then send this to
the master branch.

1. Convention
Ever since discovering C# properties, I've been abusing them when a method would
do just as well. Replaced those. Also started getting bothered by the layout of
fields, properties, and methods, so I learned what the "proper" ordering for
these things in a class are. It's still kinda strange not to have header files
to further clean things up, but ah well.

2. Movement refactoring
This is a fairly significant change in terms of quantity at least. First, I do a
way with Oculus vs Steam. Let's just deal with the speed changes if I ever get
to test it (or some sort of settings menu to adjust speed of movement in the
future). Second, NO MORE STEAMVR COMPONENTS for movement. It's cluttering up the
editor inspector. I'd rather treat it like other input types, solely in code. So
that's what I've done. As one can imagine, this reduces the amount of
unnecessary code in PlayerMove quite nicely.

As an aside, all classes starting with "Player" are now the central interfaces
for user input. That way I can sanely manage all the various types of input a
user can use (keyboard, OVR, Steam).

3. Audio layout
Audio sources are now separate objects that are children of the player. A
separate source has been made for each type of audio (narrator, bgm, and player
recording), so that we can keep the volume among the 3 types consistent. The
real issue is basically that recordings are too soft and narrator clips/bgm are
much louder in comparison. If I bring the latter down to the former's level, I
can at least make it so upping the system volume is a generally pleasant
solution.

4. Recording
Recording now works again, so we can get past that part of the Narrator
introduction! Narrator required some bug fixing to get it to work right again.
SteamVR input action set has been updated to include the rest of the missing
buttons (record, play, and next).

That should about cover everything. Next up is general recording/playback
indicators in the GUI, then the rest of the game (the npcs and their various
lessons).