Recording: General life improvement by adding GUI card
As the title of the commit says, this change is focused mainly on adding a new
quality-of-life idea: a set of icons directly above the controller that indicate
if you're recording or playing your personal audio. The icons will blink if
active and be static otherwise. It's not perfect, but it's very cool and fun to
see in the hand. Only tested on right hand for now, but will try on left.