Updated settings page with taiwan flag theme; settings refactor start
In trying to make the front page pretty, I figured out that the flag of Taiwan
has a pretty good color scheme I could take advantage of. I made some decent
backgrounds for the main text areas and even included the Taiwan sun in the
title of the game.

Now I'm workingon the settings scripts and refactoring them. I really dislike
the layout of the dropdown settings and I realized some of it was pretty ugly.
Also, it's nicer to give the player some default settings so they can just go,
right? A settings menu that forces them to choose options first is probably
pretty lame. Not done yet, but will continue here.