Settings scripts adjusted/refactored further. Microphone handling.
The main deal of this commit is what happens if the microphone gets unplugged
during gameplay? If the user is playing, they get switched to the next available
mic or booted back to the settings screen if none exists. For now, you just
can't play without a mic.

The bigger code change is probably what to do in the settings menu. For this, we
update the dropdown menu. We check once per second to see if the list has
changed. There's a GUI hiccup when this happens, but it's not so bad and it's
understandable given you unplugged your mic.  Refactoring was a big deal because
the functions felt SO messy. I think it's better now.