First test for settings menu: replace player style
When you press escape, the player is toggled between Oculus and No VR. I was
worried the change would be noticeable to the player, but looking at the game
view, it's pretty quick. Plus, if someone is really switching between headset
and desktop, it probably doesn't need to be too quick.