New AudioSourceWrapper and PlayPauseBehaviour classes
The audio changes now work pretty well when I press escape. It's pretty messy to
check a static bool named Paused though, so I had this idea to make all
"pausable" classes inherit from a class that would force good separation of code
into "playing" action and "pausing" action. That turned out to be a little over
the top however, as currently only VolumeController uses it. But
AudioSourceWrapper on the other hand is great. It's the reason Dialogue,
Narrator, IntroductionNarrator, and ObjectFocus don't need PlayPauseBehavior.
The AudioSource was really lacking in properties or callbacks that would just
tell me when a clip had actually ended or was paused. Finally creating this
class eliminated code duplication in those respective classes and made the
current state of the audio much clearer. With all this refactoring complete, the
actual test of pressing escape during conversations and the introduction speech
work as expected.