Recording handles pause, Percentage display in settings, Other improvements
So for recording, I've decided that opening the settings menu just cuts the
recording short. That's an acceptable action. I thought about having the
recording "pause" and continue after the menu ends. But no one has a strong
dependency on recording for state right now and it might not be intuitive to the
user to have the recording cut out midway through. Plus I'd have to implement an
array of audio clips or something since the Microphone class doesn't have a
built-in way to "continue" an audio clip or pause a recording.

Percentage display is added to give the user a number to look at when adjusting
volume. It feels good.

To make placement of objects easier, I've moved up the environment so the city
is at 0. Player objects may be in odd places because I haven't tested the new
player positions with VR headsets yet. Especially Oculus, I don't have that one.

Strafe left and right using left/right arrow keys is added now. Works as
expected.

Rotation vertically sometimes bugged out and allowed the player to rotate upside
down. I haven't proved it, but I blame eulerAngles.x because the documentation
explains that the quaternion isn't guaranteed to represent eulerAngles the same
way every time (there are multiple representations). A quick fix here is to keep
track of x euler angles ourselves. It works as expected.