Fixed broken Narration components, created json file for audio clips
It's been too long that I haven't documented which audio clips go where. Finally
did that. Also fixed the narration components because I think I renamed those in VS
code instead of Unity3D.

A future goal should be to create the arrays and Dialogue/Narration objects FROM the
json file. For now, I'm not doing that, but that's the ideal.